separate controller indexing owners
